+++ packages/block-editor/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Bug Fix
+
+-   `FontSizePicker`: Update fluid utils so that only string, floats and integers are treated as valid font sizes for the purposes of fluid typography.([#44847](https://github.com/WordPress/gutenberg/pull/44847))
+
+## 10.2.0 (2022-10-05)
+
+## 10.1.0 (2022-09-21)
+
 ## 10.0.0 (2022-09-13)
 
 ### Breaking change